Create separate GraphEdge objects for each direction
Sharing the same GraphEdge object for both directions was troublesome
in that we had to check the direction all the time whenever we want to
see if vehicles can interact.

Having distinct objects for both directions allow us to have statistics
for each directions separately.

Another benefit of dividing the objects is that we could easily make
one-way road by just creating half of the objects.